requirements.txt
django~=4.2.28
djangorestframework~=3.16.1
psycopg2-binary
drf-spectacular
django-filter
